validate last empty point on mvrv calculation
